requirements.txt
numpy
pandas
networkx
cosine-similarity
tqdm
mysql-connector-python
matplotlib
websocket-client
websockets
scikit-learn
more-itertools
sentence-transformers
k-means-constrained